requirements.txt
flask==0.8
gunicorn
PIL
pycrypto
